Add 64-bit VM support with MSVC x64 SSE assembly routines.

+++ code/qcommon/vm.c
@@ -590,11 +590,11 @@ intptr_t QDECL VM_Call(vm_t* vm, int callnum, ...) {
     // if we have a dll loaded, call it directly
     if (vm->entryPoint) {
         // rcg010207 -  see dissertation at top of VM_DllSyscall() in this file.
-        int args[MAX_VMMAIN_ARGS - 1];
+        intptr_t args[MAX_VMMAIN_ARGS - 1];
         va_list ap;
         va_start(ap, callnum);
         for (i = 0; i < ARRAY_LEN(args); i++) {
-            args[i] = va_arg(ap, int);
+            args[i] = va_arg(ap, intptr_t);
         }
         va_end(ap);
